requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
httpx==0.28.1
passlib==1.7.4
python-multipart==0.0.20

# Dev / Test
pytest==8.3.4
pytest-asyncio==0.25.0
